Add source file rename mechanism on Page Rename Handler

+++ Backend/handlers/page_handler.py
@@ -15,6 +15,7 @@ from ..models.page_model import (
 from ..utilities.page_utils import rename_page_references_in_str
 
 def handle_get_all_pages(page_path: Path):
+    # TODO add file metadata implementation
     file_names = os.listdir(str(page_path))
     pages = []
     for file_name in file_names:
@@ -54,11 +55,20 @@ def handle_update_page(page_path: Path, page_info: PageWithContentWithoutMetaDat
     return True
 
 def handle_page_rename(page_path: Path, rename_info: PageRenameInfo):
+    # Update the references to the source Page
     for reference in rename_info.references_to_update:
         # TODO run these operations in parallel
         path = page_path / reference.page_name # TODO handle folders - wayyyyy later down the road
         with open(path, 'r') as f:
             content = f.read()
         replaced_references = rename_page_references_in_str(rename_info.old_name, rename_info.new_name, content)
-        with open(path, 'w') as f:
-            f.write(replaced_references)
+        if replaced_references != content: # Only write to the file if something changed - prevents weird file metadata changes
+            with open(path, 'w') as f:
+                f.write(replaced_references)
+
+    # Rename the source Page file
+    old_file_name = f'{page_path / rename_info.old_name}.md'
+    new_file_name = f'{page_path / rename_info.new_name}.md'
+    os.rename(old_file_name, new_file_name) # TODO should we throw a specific error if this fails?
+
+    return {'msg': f'Updated {len(rename_info.references_to_update)} reference(s) and changed {rename_info.old_name} to {rename_info.new_name}'}
